+++ packages/modules/aave-v3/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @infinit-xyz/aave-v3
 
+## 0.0.0-alpha.23
+
+### Patch Changes
+
+- [#111](https://github.com/infinit-xyz/infinit-library/pull/111)
+  [`7ed8c28`](https://github.com/infinit-xyz/infinit-library/commit/7ed8c2885e96b8d9b34e3fb0f07dccd6edb85734) Thanks
+  [@mintcnn](https://github.com/mintcnn)! - Bump core version
+
+## 0.0.0-alpha.22
+
+### Patch Changes
+
+- [`4967948`](https://github.com/infinit-xyz/infinit-library/commit/4967948a32e8519387d53e3ec34242de98eded11) Thanks
+  [@melvinmcrn](https://github.com/melvinmcrn)! - bump core version
+
+- Updated dependencies [[`b9624e1`](https://github.com/infinit-xyz/infinit-library/commit/b9624e1797c500b33ffe4695013ee4d07191c737)]:
+  - @infinit-xyz/core@0.1.0-alpha.8
+
 ## 0.0.0-alpha.21
 
 ### Minor Changes